\name{STEPCAM-package}
\alias{STEPCAM-package}
\docType{package}
\title{
ABC-SMC inference of the STEPCAM model 
}
\description{
ABC-SMC inference of the relative contribution of stochasticy (random removal), filtering and competition (limiting similarity) through the use of stepwise commmunity assembly.

}
\details{
\tabular{ll}{
Package: \tab STEPCAM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-08\cr
License: \tab GPL 2.0\cr
}
 The function STEPCAM_ABC performs an ABC-SMC analysis of the STEPCAM model on user provided data,
 the functions plotSMC, plotSTEPCAM and TernPlot provide functions to plot the output generated by the ABC-SMC analysis.
}
\author{
Thijs Janzen & Fons van der Plas

Maintainer: Thijs Janzen <thijsjanzen@gmail.com>
}
\references{
F. van der Plas, T. Janzen, A. Ordonez, W. Fokkema, J. Reinders, R.S. Etienne & H. Olff, 2014, A new modeling approach quantifies the relative importance of different community assembly processes, in prep.

}
\keyword{ package, STEPCAM, community assembly, functional diversity, stochasticity, limiting similarity }

