\name{plotSTEPCAM}
\alias{plotSTEPCAM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Plot the output generated by the STEPCAM_ABC function
}
\description{
  This function plots the posterior distribution inferred by the STEPCAM_ABC function for the three community assembly parameters dispersal assembly, habitat filtering and limiting similarity.
}
\usage{
plotSTEPCAM(output)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{output}{
  The list generated by the function STEPCAM_ABC as output, this list contains the following elements: \cr
    \itemize{
            - \item DA a vector containing the posterior distribution for the parameter dispersal assembly \cr
            - \item HF a vector containing the posterior distribution for the parameter habitat filtering \cr
            - \item LS a vector containing the posterior distribution for the parameter limiting similarity \cr
     }
  }
}

\author{Thijs Janzen}

\examples{
\dontrun{
Artificial.Data <- generate.Artificial.Data(n_species = 40, n_traits = 3,
                        n_communities = 5, occurence_distribution = 0.2,
                        average_richness = 0.5, sd_richness = 0.2,
                        mechanism_random=FALSE)
O <- STEPCAM_ABC(Artificial.Data$abundances,Artificial.Data$traits,
                 numParticles=10,n_traits=3,plot_number=1,stopRate=0.8);
plotSTEPCAM(O);
}
}
