\name{STEPCAM-package}
\alias{STEPCAM-package}
\docType{package}
\title{
ABC-SMC Inference of STEPCAM
}
\description{
ABC-SMC inference of the relative contribution of dispersal assembly, filtering and limiting similarity through the use of stepwise commmunity assembly.
\cr \cr
Updates \cr
Version 1.2 \cr
  - STEPCAM can now be found on GitHub: https://github.com/thijsjanzen/STEPCAM \cr
  - Merged STEPCAM code with the supplementary code of Hauffe et al. 2016 \cr
  - Added two vignettes \cr
  - Added a large number of code tests \cr
Version 1.1.2 \cr
	- updated the prior - now all three processes have a mean prior value of 1/3, and a log-normal kind shape in [0,1]
Version 1.1.1 \cr
	- fixed a bug in the ABC-SMC code leading to the error "Error in if (fit < threshold)  : missing value where TRUE/FALSE needed" \cr
Version 1.1 \cr
  - calculation of the diversity indices is plit up into two parts: calculating the ordination axes, and the actual calculation of the indices - this allows for a speed up of the code by moving the calculation of the ordination axes outside the SMC algorithm. Many thanks to Torsten Hauffe for this improvement! \cr
  - Corrected scaling of the dbFD function - it now always treats abundances as presence/absence data, much thanks to Torsten Hauffe for pointing this out! \cr
  - updated references \cr
}
\details{
\tabular{ll}{
Package: \tab STEPCAM\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2016-09-20\cr
License: \tab GPL 2.0\cr
}
 The function STEPCAM_ABC performs an ABC-SMC analysis of the STEPCAM model on user provided data,
 the functions plotSMC, plotSTEPCAM and TernPlot provide functions to plot the output generated by the ABC-SMC analysis.
}
\author{
Thijs Janzen & Fons van der Plas

Maintainer: Thijs Janzen <thijsjanzen@gmail.com>
}
\references{
Fons van der Plas, Thijs Janzen, Alejandro Ordonez, Wimke Fokkema, Josephine Reinders, Rampal S. Etienne, and Han Olff 2015. A new modeling approach estimates the relative importance of different community assembly processes. Ecology 96:1502-1515.

Hauffe, Torsten, Christian Albrecht, and Thomas Wilke. Assembly processes of gastropod community change with horizontal and vertical zonation in ancient Lake Ohrid: a metacommunity speciation perspective. Biogeosciences 13.10 (2016): 2901-2911.

}
\keyword{STEPCAM}
\keyword{community assembly}
\keyword{functional diversity}
\keyword{stochasticity}
\keyword{limiting similarity}